Reorganizing the zerocash tree.
