--- src/zerocash/.travis.yml
@@ -1,35 +1,0 @@
-os: linux
-
-env:
-  global:
-    - ZCTRAVIS_BOOST_OPTS="threading=multi link=static -sNO_BZIP2=1 -sNO_ZLIB=1 address-model=64"
-    - ZCTRAVIS_BOOST_LIBRARIES="chrono,filesystem,program_options,system,thread,test"
-    - ZCTRAVIS_SNARK_FLAGS="-j2 CURVE=ALT_BN128 NO_PROCPS=1 NO_GTEST=1 NO_DOCS=1 STATIC=1"
-    - ZCTRAVIS_LIBZEROCASH_CXXFLAGS="-w -DCURVE_ALT_BN128 -DBOOST_SPIRIT_THREADSAFE -DHAVE_BUILD_INFO -D__STDC_FORMAT_MACROS -U_FORTIFY_SOURCE -D_FORTIFY_SOURCE=2 -std=c++11 -pipe -O2 -O0 -g -Wstack-protector -fstack-protector-all -fPIE -fvisibility=hidden -fPIC"
-    - ZCTRAVIS_LIBZEROCASH_FLAGS="-j2 NO_PROCPS=1 MINDEPS=1 LINK_RT=1"
-
-before_install:
-  - sudo add-apt-repository ppa:ubuntu-toolchain-r/test -y # For g++-4.9
-  - sudo apt-get update -qq
-  - sudo apt-get install -qq g++-4.9
-  - sudo update-alternatives --install /usr/bin/g++ g++ /usr/bin/g++-4.9 90
-  # Install dependencies
-  - sudo apt-get install cmake libgmp-dev libssl-dev
-  # We still need boost, let's get it.
-  - wget http://downloads.sourceforge.net/project/boost/boost/1.57.0/boost_1_57_0.tar.bz2
-  - echo "910c8c022a33ccec7f088bd65d4f14b466588dda94ba2124e78b8c57db264967  boost_1_57_0.tar.bz2" | sha256sum -c
-  - tar -xjf boost_*.tar.bz2
-  - cd boost_1_57_0
-  - ./bootstrap.sh --without-icu --with-libraries="$ZCTRAVIS_BOOST_LIBRARIES"
-  - ./b2 -d2 -j2 -d1 $ZCTRAVIS_BOOST_OPTS
-  - sudo ./b2 -d0 -j2 $ZCTRAVIS_BOOST_OPTS install
-  - cd ..
-  # We still need crypto++, let's get it.
-  - ./get-cryptopp
-
-script:
-  - LIBSNARK_FLAGS="$ZCTRAVIS_SNARK_FLAGS" ./get-libsnark
-  - CXXFLAGS="$ZCTRAVIS_LIBZEROCASH_CXXFLAGS" make $ZCTRAVIS_LIBZEROCASH_FLAGS
-#  - ./zerocash_pour_ppzksnark/tests/test_zerocash_pour_ppzksnark
-#  - ./tests/merkleTest
-#  - ./tests/zerocashTest
