+++ src/zerocash/Coin.cpp
