+++ src/zerocash/Coin.h
