+++ src/zerocash/CoinCommitment.cpp
