+++ src/zerocash/CoinCommitment.h
