+++ src/zerocash/IncrementalMerkleTree.cpp
