+++ src/zerocash/MintTransaction.cpp
