+++ src/zerocash/MintTransaction.h
