+++ src/zerocash/ZerocashParams.h
@@ -15,7 +15,7 @@
 #include "Zerocash.h"
 #include "libsnark/common/default_types/r1cs_ppzksnark_pp.hpp"
 #include "libsnark/zk_proof_systems/ppzksnark/r1cs_ppzksnark/r1cs_ppzksnark.hpp"
-#include "zerocash_pour_ppzksnark/zerocash_pour_ppzksnark.hpp"
+#include "zerocash_pour_ppzksnark.hpp"
 
 namespace libzerocash {
