+++ src/zerocash/utils/util.cpp
